requirements.txt
langchain>=0.2
langgraph>=0.1
openai>=1.0
qdrant-client>=1.9
sentence-transformers>=3.0
unstructured[pdf,docx]>=0.14
pdfplumber>=0.11
pandas>=2.2
duckdb>=0.10
fastapi>=0.111
uvicorn>=0.30
python-dotenv>=1.0
jinja2>=3.1
pydantic-settings>=2.0